warning

+++ src/gsCore/gsTemplateTools.h
@@ -239,7 +239,7 @@ bool less(T1 t1, T2 t2)
         return true;
     else if (t2 < 0 && t1 >= 0)
         return false;
-    return (t1 < t2);
+    return (static_cast<size_t>(t1) < static_cast<size_t>(t2));
 }
 
 /// Compares two (integer) numbers of even different type.
@@ -256,7 +256,7 @@ bool less_equal(T1 t1, T2 t2)
     else if (t2 < 0 && t1 >= 0)
         return false;
 
-    return (t1 <= t2);
+    return (static_cast<size_t>(t1) <= static_cast<size_t>(t2));
 }
 
 /// Compares two (integer) numbers of even different type.
@@ -291,7 +291,7 @@ bool equal(T1 t1, T2 t2)
     if ((t1 < 0 && t2 >= 0) || (t2 < 0 && t1 >= 0))
         return false;
 
-    return (t1 == t2);
+    return (static_cast<size_t>(t1) == static_cast<size_t>(t2));
 }
 
 /*
